requirements.txt
flask==3.0.0
torch==2.1.0
torchvision==0.16.0
numpy==1.26.0
Pillow==10.1.0
opencv-python-headless==4.8.1.78
gunicorn==21.2.0
